Add comments and rename to commons
